distilldsm_rak/requirements.txt
torch==1.12.1
torchvision==0.13.1
torchmetrics
pydicom
scikit-learn
tqdm
pandas
requests==2.26.0
openvino-dev[onnx]==2021.4.2
onnxruntime==1.8.1
numpy==1.19.2
matplotlib==3.5.2
wget
tqdm
pytest
